Removed old images that are no longer in use together with the HUDRenderer. The nifty-HUD is now finished (well, at least usable) and is set to standard now.

+++ Oberien/src/event/PlayerStatsAdapter.java
@@ -19,19 +19,19 @@ public class PlayerStatsAdapter {
 		listener.remove(l);
 	}
 
-	public void moneyChanged(int metal) {
+	public void moneyChanged(float metal) {
 		for (PlayerStatsListener l : listener) {
 			l.moneyChanged(metal);
 		}
 	}
 
-	public void energyChanged(int energy) {
+	public void energyChanged(float energy) {
 		for (PlayerStatsListener l : listener) {
 			l.energyChanged(energy);
 		}
 	}
 
-	public void populationChanged(int population) {
+	public void populationChanged(float population) {
 		for (PlayerStatsListener l : listener) {
 			l.populationChanged(population);
 		}
